简化了SRB的头文件 现在usb不会包含Libusb头文件 现在只需要包含Master和相关节点的头文件即可调用节点 但是设计cluster的操作仍然需要调用对应的cluster头文件

+++ support/PerformanceAnalyzer.cpp
@@ -56,7 +56,7 @@ int PerformanceTimer::beginCheck(){
 int PerformanceTimer::endCheck(){
 	long long int  end = getTimesUs();
 	_is_checking = false;
-	_last_time_cost = (end - check_buf);
+	_last_time_cost = (int)(end - check_buf);
 	return done;
 }
